+++ api/rocparser.h
@@ -277,10 +277,10 @@ extern rocDecStatus ROCDECAPI rocDecParseVideoData(RocdecVideoParser parser_hand
 
 /************************************************************************************************/
 //! \ingroup group_rocparser
-//! \fn rocDecStatus ROCDECAPI rocDecParserReleaseFrame(RocdecVideoParser parser_handle, int pic_idx)
-//! Release frame with index pic_idx from parser's buffer pool and mark it for reuse 
+//! \fn rocDecStatus ROCDECAPI rocDecParserMarkForReuse(RocdecVideoParser parser_handle, int pic_idx)
+//! Mark frame with index pic_idx in parser's buffer pool for reuse (means the frame has been consumed) 
 /************************************************************************************************/
-extern rocDecStatus ROCDECAPI rocDecParserReleaseFrame(RocdecVideoParser parser_handle, int pic_idx);
+extern rocDecStatus ROCDECAPI rocDecParserMarkFrameForReuse(RocdecVideoParser parser_handle, int pic_idx);
 
 /************************************************************************************************/
 //! \ingroup group_rocparser