+++ src/rocdecode/roc_decoder.h
@@ -62,7 +62,6 @@ public:
     rocDecStatus GetDecodeStatus(int pic_idx, RocdecDecodeStatus* decode_status);
     rocDecStatus ReconfigureDecoder(RocdecReconfigureDecoderInfo *reconfig_params);
     rocDecStatus GetVideoFrame(int pic_idx, void *dev_mem_ptr[3], uint32_t horizontal_pitch[3], RocdecProcParams *vid_postproc_params);
-    rocDecStatus ReleaseVideoFrame(RocdecVideoParser parser_hdl, int pic_idx);
 
 private:
     rocDecStatus InitHIP(int device_id);